+++ src/components/stable/gux-modal/gux-modal.types.ts
@@ -1,1 +1,1 @@
-export type GuxModalSize = 'small' | 'medium' | 'large';
+export type GuxModalSize = 'small' | 'medium' | 'large' | 'dynamic';